Merge pull request #3 from yujinS0/develop
우편함, 출석부 완료 & 피드백 수정 & Friend 관련 문서 작업 완료

+++ GameServer/DTO/Mail.cs
@@ -32,7 +32,7 @@ public class MailDetailResponse
     public Int64 MailId { get; set; }
     public string Title { get; set; }
     public string Content { get; set; }
-    public Int64 ItemCode { get; set; }
+    public int ItemCode { get; set; }
     public int ItemCnt { get; set; }
     public DateTime? SendDate { get; set; }
     public DateTime? ExpireDate { get; set; }
